domains/study/slide_to_vid/requirements.txt
boto3
pdf2image
pytesseract
pydub
rich
google-generativeai
transformers

moviepy==1.0.3
imageio==2.34.0
imageio-ffmpeg==0.4.9
numpy
decorator
tqdm

git+https://github.com/nari-labs/dia.git
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
flatten-dict==0.4.2
julius==0.2.7
ffmpy==0.3.1
randomname==0.2.1
argbind==0.3.9
